Added methods that do nothing

+++ shop_inventory/product.rb
@@ -1,3 +1,6 @@
+require 'fileutils'
+require 'tempfile'
+
 class Product
   def initialize
     @id = @name = @price = @count = @company = nil
@@ -6,8 +9,9 @@ class Product
   attr_accessor :id, :name, :price, :count, :company
 
   def add_product
+    id = get_max_id.succ
     File.open('inventory', 'a') do |file|
-      file.puts "id: #{get_max_id.succ}"
+      file.puts "id: #{id}"
       file.puts "name: #{name}"
       file.puts "price: #{price}"
       file.puts "count: #{count}"
@@ -30,15 +34,36 @@ class Product
 
   def get_max_id
     max = 0
-    reuturn max if !File.exists?('inventory')
+    return max if !File.exists?('inventory')
     File.open('inventory', 'r') do |file|
-      line = file.readline || ""
+      line = file.readline
       if line.match(/^id: (.*)/)
-        temp = $1.to_i
-        max = temp if temp > max
+        temp_id = $1.to_i
+        max = temp_id if temp_id > max
+      else
+        4.times { file.readline }
       end
     end
     max
   end
-  
+
+  def self.remove_product(id)
+    return if !File.exists?('inventory')
+    temp = Tempfile.new
+    File.open('inventory', 'r') do |file|
+      line = file.readline
+      line.match(/^id: (.*)/)
+      temp_id = $1.to_i
+      if temp_id != id
+        temp << line
+        temp << file.readline
+        temp << file.readline
+        temp << file.readline
+        temp << file.readline
+      else
+        4.times { file.readline }
+      end
+    end
+    FileUtils.mv(temp.path, 'inventory')
+  end
 end
